pyproject.toml
[build-system]
requires = ["setuptools>=45", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["time_guardian"]

[project]
name = "time-guardian"
authors = [
    {name = "Bryce Drennan"},
]
description = "Python-based screen activity tracker that captures images every 5 seconds for AI-powered computer usage analysis."
readme = "README.md"
requires-python = ">=3.10"
dynamic = ["version"]
dependencies = [
    "pillow",
    "typer",
    "requests",
    "openai",
    "numpy",
    "mss",
    "schedule",
]

[project.optional-dependencies]
dev = [
    "coverage",
    "mypy",
    "pytest",
    "pytest-randomly",
    "pytest-sugar",
    "ruff",
    "wheel",
]

[project.urls]
Documentation = "https://github.com/brycedrennan/time-guardian"
Source = "https://github.com/brycedrennan/time-guardian"

[tool.setuptools_scm]
write_to = "time_guardian/_version.py"

[tool.mypy]
ignore_missing_imports = true

[tool.ruff]
extend-exclude = []
line-length = 120
output-format = "concise"

[tool.ruff.lint]
extend-ignore = [
    "E501", "G004", "ISC001",  "PT011", "RET504", "SIM114", "TRY003", "TRY400", "TRY401", "RUF012", "RUF100"
]
extend-select = [
    "I", "E", "W", "UP", "ASYNC", "BLE", "A001", "A002",
    "C4", "DTZ", "T10", "EM", "ISC", "ICN", "G", "PIE", "PT",
    "Q", "SIM", "TID", "TCH", "PLC", "PLE", "TRY", "RUF"
]

[tool.ruff.lint.isort]
combine-as-imports = true

[tool.ruff.lint.flake8-errmsg]
max-string-length = 50

[tool.pytest.ini_options]
addopts = "-s --tb=native -v --durations=10"
norecursedirs = ["build", "dist"]

[project.scripts]
time-guardian = "time_guardian.cli:app"
